+++ test/units/static/JBController/TestMintTokensOfUnits.sol
@@ -48,9 +48,9 @@ contract TestMintTokensOfUnits_Local is JBControllerSetup {
 
         JBRuleset memory data = JBRuleset({
             cycleNumber: 1,
-            id: block.timestamp,
+            id: uint48(block.timestamp),
             basedOnId: 0,
-            start: block.timestamp,
+            start: uint48(block.timestamp),
             duration: 8000,
             weight: 5000,
             decayRate: 0,
