+++ test/units/static/JBController/TestSetTokenFor.sol
@@ -53,9 +53,9 @@ contract TestSetTokenFor_Local is JBControllerSetup {
 
         JBRuleset memory ruleset = JBRuleset({
             cycleNumber: 1,
-            id: block.timestamp,
+            id: uint48(block.timestamp),
             basedOnId: 0,
-            start: block.timestamp,
+            start: uint48(block.timestamp),
             duration: 8000,
             weight: 5000,
             decayRate: 0,
@@ -122,9 +122,9 @@ contract TestSetTokenFor_Local is JBControllerSetup {
 
         JBRuleset memory upcoming = JBRuleset({
             cycleNumber: 1,
-            id: block.timestamp,
+            id: uint48(block.timestamp),
             basedOnId: 0,
-            start: block.timestamp,
+            start: uint48(block.timestamp),
             duration: 8000,
             weight: 5000,
             decayRate: 0,
@@ -179,9 +179,9 @@ contract TestSetTokenFor_Local is JBControllerSetup {
 
         JBRuleset memory ruleset = JBRuleset({
             cycleNumber: 1,
-            id: block.timestamp,
+            id: uint48(block.timestamp),
             basedOnId: 0,
-            start: block.timestamp,
+            start: uint48(block.timestamp),
             duration: 8000,
             weight: 5000,
             decayRate: 0,
