+++ test/units/static/JBController/TestTransferCreditsFrom.sol
@@ -59,9 +59,9 @@ contract TestTransferCreditsFrom_Local is JBControllerSetup {
         // mocked ruleset that allows credit transfer
         JBRuleset memory _returnedRuleset = JBRuleset({
             cycleNumber: 1,
-            id: block.timestamp,
+            id: uint48(block.timestamp),
             basedOnId: 0,
-            start: block.timestamp,
+            start: uint48(block.timestamp),
             duration: 100,
             weight: 1e18,
             decayRate: 0,
@@ -115,9 +115,9 @@ contract TestTransferCreditsFrom_Local is JBControllerSetup {
         // mocked ruleset that allows credit transfer
         JBRuleset memory _returnedRuleset = JBRuleset({
             cycleNumber: 1,
-            id: block.timestamp,
+            id: uint48(block.timestamp),
             basedOnId: 0,
-            start: block.timestamp,
+            start: uint48(block.timestamp),
             duration: 100,
             weight: 1e18,
             decayRate: 0,
